* creatures can now be spawned on boats using the transport_creatures table. objects and stuff will follow soon. *note* this currently is only the visual of the creature, movement etc will not work. (but they will move with the boat of course) * fixed transport bug where it would crash mac clients when changing maps or loading on a transport * some tweaks to transport timing system, this still needs a *lot* of work which will come soon. :P * improved performance of transport retrival from objectmgr * corrected a9 flag for boats
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2553 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/ObjectMgr.h
@@ -297,7 +297,7 @@ public:
 	typedef HM_NAMESPACE::hash_map<uint32, Guild*>                      GuildMap;
 	typedef HM_NAMESPACE::hash_map<uint32, skilllinespell*>             SLMap;
 	typedef HM_NAMESPACE::hash_map<uint32, std::vector<CreatureItem>*>  VendorMap;
-    typedef HM_NAMESPACE::hash_map<uint64, Transporter*>                TransportMap;
+    typedef HM_NAMESPACE::hash_map<uint32, Transporter*>                TransportMap;
 	typedef HM_NAMESPACE::hash_map<uint32, Trainer*>                    TrainerMap;
 	typedef HM_NAMESPACE::hash_map<uint32, std::vector<TrainerSpell*> > TrainerSpellMap;
     typedef HM_NAMESPACE::hash_map<uint32, ReputationModifier*>         ReputationModMap;
@@ -504,7 +504,7 @@ public:
 		return r;
 	}
 
-	Transporter * GetTransporter(uint64 guid);
+	Transporter * GetTransporter(uint32 guid);
 	Transporter * GetTransporterByEntry(uint32 entry);
 
 	Charter * CreateCharter(uint32 LeaderGuid, CharterTypes Type);
